requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
cssselect==1.2.0
docopt==0.6.2
idna==3.4
lxml==4.9.3
readability-lxml==0.8.1
requests==2.31.0
soupsieve==2.4.1
tqdm==4.66.1
urllib3==2.0.4
